+++ examples/dao/test/happy-path.js
@@ -2,7 +2,7 @@ const { AccountStore } = require('@algo-builder/runtime');
 const { types, parsing } = require('@algo-builder/web');
 const { assert } = require('chai');
 const { Context, initialBalance, minSupport, deposit } = require('./common');
-const { ProposalType, Vote } = require('../scripts/run/common/common');
+const { ProposalType, Vote, DAOActions } = require('../scripts/run/common/common');
 const {
   mkProposalTx, mkDepositVoteTokenTx, votingStart,
   votingEnd, executeBefore, mkWithdrawVoteDepositTx,
@@ -11,7 +11,7 @@ const {
 
 describe('DAO - Happy Paths', function () {
   let master, creator, proposer, voterA, voterB;
-  let depositLsigAcc, daoFundLsigAcc, proposalLsigAcc;
+  let daoFundLsigAcc, proposalLsigAcc;
   let ctx;
 
   function setUpCtx () {
@@ -20,13 +20,12 @@ describe('DAO - Happy Paths', function () {
     proposer = new AccountStore(initialBalance);
     voterA = new AccountStore(initialBalance);
     voterB = new AccountStore(initialBalance);
-    depositLsigAcc = new AccountStore(initialBalance); // runtime.account of depositLsig.address()
     daoFundLsigAcc = new AccountStore(initialBalance);
     proposalLsigAcc = new AccountStore(initialBalance);
 
     ctx = new Context(
       master, creator, proposer, voterA, voterB,
-      depositLsigAcc, daoFundLsigAcc, proposalLsigAcc
+      daoFundLsigAcc, proposalLsigAcc
     );
   }
 
@@ -51,11 +50,11 @@ describe('DAO - Happy Paths', function () {
     ctx.addProposal();
 
     // deposit & register yes votes (by A)
-    ctx.depositVoteToken(ctx.voterA, ctx.depositLsig, 5);
+    ctx.depositVoteToken(ctx.voterA, 5);
     ctx.vote(ctx.voterA, Vote.YES, ctx.proposalLsigAcc);
 
     // deposit & register yes votes (by B)
-    ctx.depositVoteToken(ctx.voterB, ctx.depositLsig, 7);
+    ctx.depositVoteToken(ctx.voterB, 7);
     ctx.vote(ctx.voterB, Vote.YES, ctx.proposalLsigAcc);
 
     // execute proposal
@@ -86,7 +85,6 @@ describe('DAO - Happy Paths', function () {
         ctx.daoAppID,
         ctx.govTokenID,
         ctx.proposer.account,
-        ctx.depositLsig,
         ctx.proposalLsig,
         ctx.daoFundLsig
       );
@@ -168,49 +166,49 @@ describe('DAO - Happy Paths', function () {
     };
 
     it('should accept token deposit', () => {
-      const beforeBal = ctx.depositLsigAcc.getAssetHolding(ctx.govTokenID).amount;
+      const beforeBal = ctx.depositAcc.getAssetHolding(ctx.govTokenID).amount;
 
-      _depositVoteToken(ctx.voterA.account, ctx.depositLsig, 6);
+      _depositVoteToken(ctx.voterA.account, 6);
 
       // verify local state
       assert.deepEqual(ctx.voterA.getLocalState(ctx.daoAppID, 'deposit'), 6n);
 
       // verify deposit
-      assert.deepEqual(ctx.depositLsigAcc.getAssetHolding(ctx.govTokenID).amount, beforeBal + 6n);
+      assert.deepEqual(ctx.depositAcc.getAssetHolding(ctx.govTokenID).amount, beforeBal + 6n);
     });
 
     it('should accept multiple token deposit by same & different accounts', () => {
-      const beforeBal = ctx.depositLsigAcc.getAssetHolding(ctx.govTokenID).amount;
+      const beforeBal = ctx.depositAcc.getAssetHolding(ctx.govTokenID).amount;
       const initialADeposit = ctx.voterA.getLocalState(ctx.daoAppID, 'deposit');
 
       // deposit 6 votes by A
-      _depositVoteToken(ctx.voterA.account, ctx.depositLsig, 6);
+      _depositVoteToken(ctx.voterA.account, 6);
 
       // verify local state & deposit
       assert.deepEqual(ctx.voterA.getLocalState(ctx.daoAppID, 'deposit'), initialADeposit + 6n);
-      assert.deepEqual(ctx.depositLsigAcc.getAssetHolding(ctx.govTokenID).amount, beforeBal + 6n);
+      assert.deepEqual(ctx.depositAcc.getAssetHolding(ctx.govTokenID).amount, beforeBal + 6n);
 
       // deposit 4 votes by A again
-      _depositVoteToken(ctx.voterA.account, ctx.depositLsig, 4);
+      _depositVoteToken(ctx.voterA.account, 4);
 
       // verify local state & deposit
       assert.deepEqual(ctx.voterA.getLocalState(ctx.daoAppID, 'deposit'), initialADeposit + 6n + 4n);
-      assert.deepEqual(ctx.depositLsigAcc.getAssetHolding(ctx.govTokenID).amount, beforeBal + 6n + 4n);
+      assert.deepEqual(ctx.depositAcc.getAssetHolding(ctx.govTokenID).amount, beforeBal + 6n + 4n);
 
       // deposit 5 votes by B
-      _depositVoteToken(ctx.voterB.account, ctx.depositLsig, 5);
+      _depositVoteToken(ctx.voterB.account, 5);
 
       // verify local state & deposit
       assert.deepEqual(ctx.voterB.getLocalState(ctx.daoAppID, 'deposit'), 5n);
-      assert.deepEqual(ctx.depositLsigAcc.getAssetHolding(ctx.govTokenID).amount, beforeBal + 6n + 4n + 5n);
+      assert.deepEqual(ctx.depositAcc.getAssetHolding(ctx.govTokenID).amount, beforeBal + 6n + 4n + 5n);
     });
 
     it('should allow token deposit by any "from" account', () => {
-      const beforeBal = ctx.depositLsigAcc.getAssetHolding(ctx.govTokenID).amount;
+      const beforeBal = ctx.depositAcc.getAssetHolding(ctx.govTokenID).amount;
       const initialADeposit = ctx.voterA.getLocalState(ctx.daoAppID, 'deposit');
 
       const depositVoteTx = mkDepositVoteTokenTx(
-        ctx.daoAppID, ctx.govTokenID, ctx.voterA.account, ctx.depositLsig, 5
+        ctx.daoAppID, ctx.govTokenID, ctx.voterA.account, 5
       );
 
       // NOTE: tokens are transferred from proposer account, but recorded in voterA account
@@ -221,7 +219,7 @@ describe('DAO - Happy Paths', function () {
 
       // verify local state & deposit
       assert.deepEqual(ctx.voterA.getLocalState(ctx.daoAppID, 'deposit'), initialADeposit + 5n);
-      assert.deepEqual(ctx.depositLsigAcc.getAssetHolding(ctx.govTokenID).amount, beforeBal + 5n);
+      assert.deepEqual(ctx.depositAcc.getAssetHolding(ctx.govTokenID).amount, beforeBal + 5n);
     });
   });
 
@@ -247,10 +245,10 @@ describe('DAO - Happy Paths', function () {
       ctx.syncAccounts();
 
       // deposit votes (by A)
-      ctx.depositVoteToken(ctx.voterA, ctx.depositLsig, 6);
+      ctx.depositVoteToken(ctx.voterA, 6);
 
       // deposit votes (by B)
-      ctx.depositVoteToken(ctx.voterB, ctx.depositLsig, 4);
+      ctx.depositVoteToken(ctx.voterB, 4);
 
       // set "now" between [votingStart, votingEnd]
       ctx.runtime.setRoundAndTimestamp(10, votingStart + Math.round((votingEnd - votingStart) / 2));
@@ -350,11 +348,11 @@ describe('DAO - Happy Paths', function () {
       ctx.syncAccounts();
 
       // deposit & register yes votes (by A)
-      ctx.depositVoteToken(ctx.voterA, ctx.depositLsig, minSupport + 1);
+      ctx.depositVoteToken(ctx.voterA, minSupport + 1);
       ctx.vote(ctx.voterA, Vote.YES, ctx.proposalLsigAcc);
 
       // deposit & register NO votes (by B)
-      ctx.depositVoteToken(ctx.voterB, ctx.depositLsig, 2);
+      ctx.depositVoteToken(ctx.voterB, 2);
       ctx.vote(ctx.voterB, Vote.NO, ctx.proposalLsigAcc);
 
       // set current time after voting over
@@ -447,7 +445,7 @@ describe('DAO - Happy Paths', function () {
     let withdrawVoteDepositTx;
     this.beforeEach(() => {
       withdrawVoteDepositTx = mkWithdrawVoteDepositTx(
-        ctx.daoAppID, ctx.govTokenID, ctx.voterA.account, ctx.depositLsig, 5
+        ctx.daoAppID, ctx.govTokenID, ctx.voterA.account, 5
       );
     });
 
@@ -464,17 +462,19 @@ describe('DAO - Happy Paths', function () {
       const beforeRecordVoterB = ctx.voterB.getLocalState(ctx.daoAppID, 'deposit');
 
       // withdrawal by A
-      ctx.executeTx([
-        { ...withdrawVoteDepositTx[0], fromAccount: ctx.voterA.account },
-        { ...withdrawVoteDepositTx[1], toAccountAddr: ctx.voterA.address, amount: 5 }
-      ]);
+      ctx.executeTx(
+        mkWithdrawVoteDepositTx(
+          ctx.daoAppID, ctx.govTokenID, ctx.voterA.account, 5
+        )
+      );
       ctx.syncAccounts();
 
       // withdrawal by B
-      ctx.executeTx([
-        { ...withdrawVoteDepositTx[0], fromAccount: ctx.voterB.account },
-        { ...withdrawVoteDepositTx[1], toAccountAddr: ctx.voterB.address, amount: 7 }
-      ]);
+      ctx.executeTx(
+        mkWithdrawVoteDepositTx(
+          ctx.daoAppID, ctx.govTokenID, ctx.voterB.account, 7
+        )
+      );
       ctx.syncAccounts();
 
       // verify sender.deposit
@@ -545,9 +545,7 @@ describe('DAO - Happy Paths', function () {
       const clearProposalParam = mkClearProposalTx(
         ctx.daoAppID,
         ctx.govTokenID,
-        ctx.depositLsig,
-        ctx.proposalLsig,
-        deposit
+        ctx.proposalLsig
       );
       ctx.executeTx(clearProposalParam);
       ctx.syncAccounts();