+++ packages/runtime/test/src/interpreter/inner-transaction.ts
@@ -19,7 +19,7 @@ export function setDummyAccInfo (acc: AccountStoreI): void {
   acc.createdApps = accInfo[0].createdApps;
 }
 
-describe("TEALv5: Inner Transactions", function () {
+describe("Inner Transactions", function () {
   const elonPk = decodeAddress(elonAddr).publicKey;
   let tealCode: string;
   let interpreter: Interpreter;
@@ -36,12 +36,15 @@ describe("TEALv5: Inner Transactions", function () {
 
   // setup application account
   const appAccAddr = getApplicationAddress(TXN_OBJ.apid);
-  const applicationAccount = new AccountStore(
-    ALGORAND_ACCOUNT_MIN_BALANCE,
-    { addr: appAccAddr, sk: new Uint8Array(0) });
+  const applicationAccount = new AccountStore(ALGORAND_ACCOUNT_MIN_BALANCE, {
+    addr: appAccAddr,
+    sk: new Uint8Array(0)
+  });
 
   const reset = (): void => {
-    while (interpreter.stack.length() !== 0) { interpreter.stack.pop(); }
+    while (interpreter.stack.length() !== 0) {
+      interpreter.stack.pop();
+    }
     interpreter.subTxn = undefined;
     interpreter.runtime.ctx.pooledApplCost = 0;
     interpreter.instructions = [];
@@ -57,10 +60,10 @@ describe("TEALv5: Inner Transactions", function () {
     });
   };
 
-  const setUpInterpreter = (): void => {
+  const setUpInterpreter = (tealVersion: number): void => {
     interpreter = new Interpreter();
     interpreter.runtime = new Runtime([elonAcc, johnAcc, bobAccount, applicationAccount]);
-    interpreter.tealVersion = MaxTEALVersion;
+    interpreter.tealVersion = tealVersion;
     reset();
   };
 
@@ -71,7 +74,7 @@ describe("TEALv5: Inner Transactions", function () {
     interpreter.execute(tealCode, ExecutionMode.APPLICATION, interpreter.runtime, 0);
   };
 
-  this.beforeAll(setUpInterpreter);
+  this.beforeAll(() => setUpInterpreter(5));
 
   describe("TestActionTypes", function () {
     it("should fail: itxn_submit without itxn_begin", function () {
@@ -106,8 +109,7 @@ describe("TEALv5: Inner Transactions", function () {
         itxn_submit
         int 1
       `;
-      assert.throws(
-        () => executeTEAL(tealCode), `unsupported type for itxn_submit`);
+      assert.throws(() => executeTEAL(tealCode), `unsupported type for itxn_submit`);
 
       tealCode = `
         itxn_begin
@@ -117,7 +119,9 @@ describe("TEALv5: Inner Transactions", function () {
         int 1
       `;
       assert.throws(
-        () => executeTEAL(tealCode), `Type does not represent 'pay', 'axfer', 'acfg' or 'afrz'`);
+        () => executeTEAL(tealCode),
+        `Type does not represent 'pay', 'axfer', 'acfg' or 'afrz'`
+      );
     });
 
     it("should fail: Type arg not a byte array", function () {
@@ -130,10 +134,7 @@ describe("TEALv5: Inner Transactions", function () {
         int 1
       `;
 
-      expectRuntimeError(
-        () => executeTEAL(tealCode),
-        RUNTIME_ERRORS.TEAL.INVALID_TYPE
-      );
+      expectRuntimeError(() => executeTEAL(tealCode), RUNTIME_ERRORS.TEAL.INVALID_TYPE);
     });
 
     it("should fail: not a uint64", function () {
@@ -146,10 +147,7 @@ describe("TEALv5: Inner Transactions", function () {
         int 1
       `;
 
-      expectRuntimeError(
-        () => executeTEAL(tealCode),
-        RUNTIME_ERRORS.TEAL.INVALID_TYPE
-      );
+      expectRuntimeError(() => executeTEAL(tealCode), RUNTIME_ERRORS.TEAL.INVALID_TYPE);
     });
 
     it("should fail: invalid types (good, but not allowed in tealv5)", function () {
@@ -162,10 +160,7 @@ describe("TEALv5: Inner Transactions", function () {
         int 1
       `;
 
-      expectRuntimeError(
-        () => executeTEAL(tealCode),
-        RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR
-      );
+      expectRuntimeError(() => executeTEAL(tealCode), RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR);
 
       tealCode = `
         itxn_begin
@@ -174,10 +169,7 @@ describe("TEALv5: Inner Transactions", function () {
         itxn_submit
         int 1
       `;
-      expectRuntimeError(
-        () => executeTEAL(tealCode),
-        RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR
-      );
+      expectRuntimeError(() => executeTEAL(tealCode), RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR);
     });
 
     it("should fail: invalid types (good, but not allowed in tealv5) for TypeEnum", function () {
@@ -190,10 +182,7 @@ describe("TEALv5: Inner Transactions", function () {
         int 1
       `;
 
-      expectRuntimeError(
-        () => executeTEAL(tealCode),
-        RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR
-      );
+      expectRuntimeError(() => executeTEAL(tealCode), RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR);
 
       tealCode = `
         itxn_begin
@@ -202,10 +191,7 @@ describe("TEALv5: Inner Transactions", function () {
         itxn_submit
         int 1
       `;
-      expectRuntimeError(
-        () => executeTEAL(tealCode),
-        RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR
-      );
+      expectRuntimeError(() => executeTEAL(tealCode), RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR);
 
       tealCode = `
         itxn_begin
@@ -214,10 +200,7 @@ describe("TEALv5: Inner Transactions", function () {
         itxn_submit
         int 1
       `;
-      expectRuntimeError(
-        () => executeTEAL(tealCode),
-        RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR
-      );
+      expectRuntimeError(() => executeTEAL(tealCode), RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR);
 
       tealCode = `
         itxn_begin
@@ -226,10 +209,7 @@ describe("TEALv5: Inner Transactions", function () {
         itxn_submit
         int 1
       `;
-      expectRuntimeError(
-        () => executeTEAL(tealCode),
-        RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR
-      );
+      expectRuntimeError(() => executeTEAL(tealCode), RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR);
     });
 
     it(`should fail: "insufficient balance" because app account is charged fee`, function () {
@@ -302,7 +282,9 @@ describe("TEALv5: Inner Transactions", function () {
     it(`should pass if app account is funded`, function () {
       // increase balance
       const acc = interpreter.runtime.ctx.state.accounts.get(appAccAddr);
-      if (acc) { acc.amount = BigInt(ALGORAND_ACCOUNT_MIN_BALANCE) * 10n; }
+      if (acc) {
+        acc.amount = BigInt(ALGORAND_ACCOUNT_MIN_BALANCE) * 10n;
+      }
 
       // default receiver is zeroAddr, amount is 0
       tealCode = `
@@ -384,9 +366,14 @@ describe("TEALv5: Inner Transactions", function () {
   });
 
   describe("TestFieldTypes", function () {
-    const ConfigAddresses = ['ConfigAssetManager', 'ConfigAssetReserve', 'ConfigAssetFreeze', 'ConfigAssetClawback'];
+    const ConfigAddresses = [
+      "ConfigAssetManager",
+      "ConfigAssetReserve",
+      "ConfigAssetFreeze",
+      "ConfigAssetClawback"
+    ];
 
-    it('should pass: teal understand 32 bytes value is address with acfg address', () => {
+    it("should pass: teal understand 32 bytes value is address with acfg address", () => {
       ConfigAddresses.forEach((configAddr) => {
         tealCode = `
         itxn_begin
@@ -518,9 +505,13 @@ describe("TEALv5: Inner Transactions", function () {
     it(`should assert sender, application balance 0 before pay`, function () {
       // set sender.balance == 0
       const elonAcc = interpreter.runtime.ctx.state.accounts.get(elonAddr);
-      if (elonAcc) { elonAcc.amount = 0n; }
+      if (elonAcc) {
+        elonAcc.amount = 0n;
+      }
       const acc = interpreter.runtime.ctx.state.accounts.get(appAccAddr);
-      if (acc) { acc.amount = 0n; }
+      if (acc) {
+        acc.amount = 0n;
+      }
       const checkBal = `
         txn Sender
         balance
@@ -554,7 +545,8 @@ describe("TEALv5: Inner Transactions", function () {
     // */
 
     it(`should fail: insufficient balance`, function () {
-      const teal = `
+      const teal =
+        `
         global CurrentApplicationAddress
         txn Accounts 1
         int 100
@@ -569,9 +561,12 @@ describe("TEALv5: Inner Transactions", function () {
     // TODO: check if this should indeed fail (if receiver account.balance < minBalance)
     it(`should pass: after increasing app account's balance`, function () {
       const acc = interpreter.runtime.ctx.state.accounts.get(appAccAddr);
-      if (acc) { acc.amount = 1000000n; }
+      if (acc) {
+        acc.amount = 1000000n;
+      }
 
-      const teal = `
+      const teal =
+        `
         global CurrentApplicationAddress
         txn Accounts 1
         int 100
@@ -597,7 +592,9 @@ describe("TEALv5: Inner Transactions", function () {
 
     it(`should test pay with closeRemTo`, function () {
       const acc = interpreter.runtime.ctx.state.accounts.get(elonAddr);
-      if (acc) { acc.amount = 1000000n; }
+      if (acc) {
+        acc.amount = 1000000n;
+      }
 
       const teal = `
         itxn_begin
@@ -643,10 +640,7 @@ describe("TEALv5: Inner Transactions", function () {
 
     it(`should fail: invalid ASA Ref`, function () {
       TXN_OBJ.apas = []; // remove foreign assets refs
-      expectRuntimeError(
-        () => executeTEAL(axfer),
-        RUNTIME_ERRORS.TEAL.INVALID_ASA_REFERENCE
-      );
+      expectRuntimeError(() => executeTEAL(axfer), RUNTIME_ERRORS.TEAL.INVALID_ASA_REFERENCE);
     });
 
     it(`should fail: ref is passed but bal == 0`, function () {
@@ -660,13 +654,12 @@ describe("TEALv5: Inner Transactions", function () {
     it(`should fail: not opted in`, function () {
       // increase app balance
       const acc = interpreter.runtime.ctx.state.accounts.get(appAccAddr);
-      if (acc) { acc.amount = 1000000n; }
+      if (acc) {
+        acc.amount = 1000000n;
+      }
 
       // sufficient bal, but not optedIn
-      expectRuntimeError(
-        () => executeTEAL(axfer),
-        RUNTIME_ERRORS.TRANSACTION.ASA_NOT_OPTIN
-      );
+      expectRuntimeError(() => executeTEAL(axfer), RUNTIME_ERRORS.TRANSACTION.ASA_NOT_OPTIN);
     });
 
     it(`should test ASA optin, asa transfer`, function () {
@@ -685,19 +678,16 @@ describe("TEALv5: Inner Transactions", function () {
       `;
 
       // does not exist
-      expectRuntimeError(
-        () => executeTEAL(optin),
-        RUNTIME_ERRORS.ASA.ASSET_NOT_FOUND
-      );
+      expectRuntimeError(() => executeTEAL(optin), RUNTIME_ERRORS.ASA.ASSET_NOT_FOUND);
 
       let assetID = 0;
       const elonAcc = interpreter.runtime.ctx.state.accounts.get(elonAddr);
       if (elonAcc) {
         assetID = interpreter.runtime.ctx.deployASADef(
-          'test-asa',
+          "test-asa",
           { total: 10, decimals: 0, unitName: "TASA" },
           elonAddr,
-          { creator: { ...elonAcc.account, name: 'elon' } }
+          { creator: { ...elonAcc.account, name: "elon" } }
         ).assetID;
       }
 
@@ -711,8 +701,12 @@ describe("TEALv5: Inner Transactions", function () {
       );
 
       // increase asa balance to 5
-      const holding = interpreter.runtime.ctx.state.accounts.get(appAccAddr)?.getAssetHolding(assetID);
-      if (holding) { holding.amount = 5n; }
+      const holding = interpreter.runtime.ctx.state.accounts
+        .get(appAccAddr)
+        ?.getAssetHolding(assetID);
+      if (holding) {
+        holding.amount = 5n;
+      }
 
       executeTEAL(axfer);
       executeTEAL(axfer);
@@ -773,18 +767,18 @@ describe("TEALv5: Inner Transactions", function () {
       if (elonAcc) {
         // in foreign-assets
         assetID1 = interpreter.runtime.ctx.deployASADef(
-          'test-asa-1',
+          "test-asa-1",
           { total: 11, decimals: 0, unitName: "TASA1" },
           elonAddr,
-          { creator: { ...elonAcc.account, name: 'elon' } }
+          { creator: { ...elonAcc.account, name: "elon" } }
         ).assetID;
 
         // not in foreign-assets
         assetID2 = interpreter.runtime.ctx.deployASADef(
-          'test-asa-2',
+          "test-asa-2",
           { total: 22, decimals: 0, unitName: "TASA2" },
           elonAddr,
-          { creator: { ...elonAcc.account, name: 'elon' } }
+          { creator: { ...elonAcc.account, name: "elon" } }
         ).assetID;
       }
 
@@ -813,10 +807,7 @@ describe("TEALv5: Inner Transactions", function () {
         ==
       `;
 
-      expectRuntimeError(
-        () => executeTEAL(teal),
-        RUNTIME_ERRORS.TEAL.INVALID_ASA_REFERENCE
-      );
+      expectRuntimeError(() => executeTEAL(teal), RUNTIME_ERRORS.TEAL.INVALID_ASA_REFERENCE);
     });
 
     it(`should fail: Sender not opted-in`, function () {
@@ -830,10 +821,7 @@ describe("TEALv5: Inner Transactions", function () {
       `;
 
       // assert failed
-      expectRuntimeError(
-        () => executeTEAL(teal),
-        RUNTIME_ERRORS.TEAL.TEAL_ENCOUNTERED_ERR
-      );
+      expectRuntimeError(() => executeTEAL(teal), RUNTIME_ERRORS.TEAL.TEAL_ENCOUNTERED_ERR);
     });
 
     it(`should fail: app account not opted in`, function () {
@@ -847,18 +835,18 @@ describe("TEALv5: Inner Transactions", function () {
       `;
 
       // assert failed
-      expectRuntimeError(
-        () => executeTEAL(teal),
-        RUNTIME_ERRORS.TEAL.TEAL_ENCOUNTERED_ERR
-      );
+      expectRuntimeError(() => executeTEAL(teal), RUNTIME_ERRORS.TEAL.TEAL_ENCOUNTERED_ERR);
     });
 
     it(`should create new holding in appAccount`, function () {
       interpreter.runtime.ctx.optIntoASA(assetID1, appAccAddr, {});
       // increase asa balance to 5
-      const holding =
-        interpreter.runtime.ctx.state.accounts.get(appAccAddr)?.getAssetHolding(assetID1);
-      if (holding) { holding.amount = 3000n; }
+      const holding = interpreter.runtime.ctx.state.accounts
+        .get(appAccAddr)
+        ?.getAssetHolding(assetID1);
+      if (holding) {
+        holding.amount = 3000n;
+      }
 
       const verifyHolding = `
         global CurrentApplicationAddress
@@ -873,23 +861,22 @@ describe("TEALv5: Inner Transactions", function () {
 
     it(`should fail: receiver not optedin`, function () {
       TXN_OBJ.apat = [Buffer.from(decodeAddress(johnAddr).publicKey)];
-      const teal = `
+      const teal =
+        `
         global CurrentApplicationAddress
         txn Accounts 1
         int 100
       ` + axfer;
 
-      expectRuntimeError(
-        () => executeTEAL(teal),
-        RUNTIME_ERRORS.TRANSACTION.ASA_NOT_OPTIN
-      );
+      expectRuntimeError(() => executeTEAL(teal), RUNTIME_ERRORS.TRANSACTION.ASA_NOT_OPTIN);
     });
 
     it(`should fail: insufficient balance`, function () {
       // optin by txn Account1
       interpreter.runtime.ctx.optIntoASA(assetID1, johnAddr, {});
 
-      const teal = `
+      const teal =
+        `
         global CurrentApplicationAddress
         txn Accounts 1
         int 100000
@@ -905,16 +892,14 @@ describe("TEALv5: Inner Transactions", function () {
       // changing to random
       TXN_OBJ.apas = [1211, 323];
 
-      const teal = `
+      const teal =
+        `
         global CurrentApplicationAddress
         txn Accounts 1
         int 100000
       ` + axfer;
 
-      expectRuntimeError(
-        () => executeTEAL(teal),
-        RUNTIME_ERRORS.TEAL.INVALID_ASA_REFERENCE
-      );
+      expectRuntimeError(() => executeTEAL(teal), RUNTIME_ERRORS.TEAL.INVALID_ASA_REFERENCE);
 
       TXN_OBJ.apas = [assetID1]; // restore
     });
@@ -935,18 +920,18 @@ describe("TEALv5: Inner Transactions", function () {
       `;
 
       // defaults to 0, which is not opted in
-      expectRuntimeError(
-        () => executeTEAL(noid),
-        RUNTIME_ERRORS.TRANSACTION.ASA_NOT_OPTIN
-      );
+      expectRuntimeError(() => executeTEAL(noid), RUNTIME_ERRORS.TRANSACTION.ASA_NOT_OPTIN);
     });
 
     it(`should pass: inner transaction axfer`, function () {
-      const teal = `
+      const teal =
+        `
         global CurrentApplicationAddress
         txn Accounts 1
         int 100
-      ` + axfer + `
+      ` +
+        axfer +
+        `
         int 1
       `;
       assert.doesNotThrow(() => executeTEAL(teal));
@@ -995,10 +980,7 @@ describe("TEALv5: Inner Transactions", function () {
         itxn_submit
       `;
 
-      expectRuntimeError(
-        () => executeTEAL(pay),
-        RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR
-      );
+      expectRuntimeError(() => executeTEAL(pay), RUNTIME_ERRORS.TEAL.ITXN_FIELD_ERR);
     });
   });
 
@@ -1017,10 +999,14 @@ describe("TEALv5: Inner Transactions", function () {
 
       assert.doesNotThrow(() => executeTEAL(pay + `int 1`));
       assert.doesNotThrow(() => executeTEAL(pay + pay + `int 1`));
-      assert.doesNotThrow(() => executeTEAL(pay + pay + pay + pay + pay + pay + pay + pay + `int 1`));
+      assert.doesNotThrow(() =>
+        executeTEAL(pay + pay + pay + pay + pay + pay + pay + pay + `int 1`)
+      );
 
       let heavyCode = ``;
-      for (let i = 0; i <= 17; ++i) { heavyCode += pay; }
+      for (let i = 0; i <= 17; ++i) {
+        heavyCode += pay;
+      }
       expectRuntimeError(
         () => executeTEAL(heavyCode),
         RUNTIME_ERRORS.GENERAL.MAX_INNER_TRANSACTIONS_EXCEEDED
@@ -1096,18 +1082,12 @@ describe("TEALv5: Inner Transactions", function () {
       `;
 
       TXN_OBJ.apas = [];
-      expectRuntimeError(
-        () => executeTEAL(freeze),
-        RUNTIME_ERRORS.TEAL.INVALID_ASA_REFERENCE
-      );
+      expectRuntimeError(() => executeTEAL(freeze), RUNTIME_ERRORS.TEAL.INVALID_ASA_REFERENCE);
 
       TXN_OBJ.apas = [createdAssetID];
       TXN_OBJ.apaa = [Buffer.from(new Uint8Array([1]))];
       // does not hold Asset
-      expectRuntimeError(
-        () => executeTEAL(freeze),
-        RUNTIME_ERRORS.TRANSACTION.ASA_NOT_OPTIN
-      );
+      expectRuntimeError(() => executeTEAL(freeze), RUNTIME_ERRORS.TRANSACTION.ASA_NOT_OPTIN);
 
       // should freeze now
       interpreter.runtime.optIntoASA(createdAssetID, johnAddr, {});
@@ -1152,10 +1132,12 @@ describe("TEALv5: Inner Transactions", function () {
       `;
 
       assert.doesNotThrow(() => executeTEAL(log));
-      const logs = interpreter.runtime.getTxReceipt(TXN_OBJ.txID)?.logs as string[];
+      const logs = interpreter.runtime.getTxReceipt(TXN_OBJ.txID)?.logs;
       assert.isDefined(logs);
 
-      for (let i = 0; i < 30; ++i) { assert.equal(logs[i], "a"); }
+      for (let i = 0; i < 30; ++i) {
+        assert.equal(logs[i], "a");
+      }
       assert.equal(logs[30], "b");
       assert.equal(logs[31], "c");
     });
@@ -1208,6 +1190,23 @@ describe("TEALv5: Inner Transactions", function () {
         () => executeTEAL(log),
         RUNTIME_ERRORS.TEAL.LOGS_LENGTH_EXCEEDED_THRESHOLD
       );
+    });
+  });
+
+  describe("Teal v6 update", function () {
+    let rekeyProgram: string;
+    this.beforeAll(() => {
+      setUpInterpreter(6);
+      rekeyProgram = ` 
+        itxn_begin
+        txn Receiver
+        itxn_field RekeyTo
+        int 1
+        return
+      `;
+    });
+    it("Should support RekeyTo", function () {
+      assert.doesNotThrow(() => executeTEAL(rekeyProgram));
     });
   });
 });